backend/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==2.2.2
pony==0.7.16
requests==2.28.1
flask-restplus==0.13.0
Werkzeug==2.2.2
Flask-Login==0.6.2
email_validator==1.3.0
Flask-WTF==1.0.1
WTForms==3.0.1
PyYAML==6.0 
flask-swagger==0.2.14
swagger-ui-py==22.7.13
python-socketio==5.7.2
Flask-SocketIO==5.3.2
pandas==1.5.2

# # Flask Packages
# Flask-Login==0.4.0
# Flask-Migrate==2.0.2
# Flask-Script==2.0.5
# Flask-SQLAlchemy==2.4.0
# Flask-WTF==0.14.2
# Flask-User==1.0.1.5

credentials~=1.1
